Refactoring files into packages for better navigation.
